+++ sonic_platform_base/sonic_xcvr/api/public/cmisVDM.py
@@ -19,23 +19,23 @@ class CmisVdmApi(XcvrApi):
         result = mantissa*10**(scale_exponent-24)
         return result
 
-    def get_VDM_page(self, page, VDM_flag_page):
+    def get_vdm_page(self, page, VDM_flag_page):
         if page not in [0x20, 0x21, 0x22, 0x23]:
             raise ValueError('Page not in VDM Descriptor range!')
-        VDM_descriptor = self.xcvr_eeprom.read_flexible(page * PAGE_SIZE + PAGE_OFFSET, PAGE_SIZE)
+        vdm_descriptor = self.xcvr_eeprom.read_raw(page * PAGE_SIZE + PAGE_OFFSET, PAGE_SIZE)
         
         # Odd Adress VDM observable type ID, real-time monitored value in Page + 4
-        VDM_typeID = VDM_descriptor[1::2]
+        vdm_typeID = vdm_descriptor[1::2]
         # Even Address
         # Bit 7-4: Threshold set ID in Page + 8, in group of 8 bytes, 16 sets/page
         # Bit 3-0: n. Monitored lane n+1 
-        VDM_lane = [(elem & 0xf) for elem in VDM_descriptor[0::2]]
-        VDM_thresholdID = [(elem>>4) for elem in VDM_descriptor[0::2]]
-        VDM_valuePage = page + 4
-        VDM_thrshPage = page + 8
-        VDM_Page_data = {}
-        VDM_TYPE_DICT = self.xcvr_eeprom.mem_map.codes['cmis_code'].VDM_TYPE
-        for index, typeID in enumerate(VDM_typeID):
+        vdm_lane = [(elem & 0xf) for elem in vdm_descriptor[0::2]]
+        VDM_thresholdID = [(elem>>4) for elem in vdm_descriptor[0::2]]
+        vdm_valuePage = page + 4
+        vdm_thrshPage = page + 8
+        vdm_Page_data = {}
+        VDM_TYPE_DICT = self.xcvr_eeprom.mem_map.codes.VDM_TYPE
+        for index, typeID in enumerate(vdm_typeID):
             if typeID not in VDM_TYPE_DICT:
                 continue
             else:
@@ -45,17 +45,17 @@ class CmisVdmApi(XcvrApi):
                 vdm_format = vdm_info_dict[1]
                 scale = vdm_info_dict[2]
 
-                vdm_value_offset = VDM_valuePage * PAGE_SIZE + PAGE_OFFSET + VDM_SIZE * index
-                vdm_high_alarm_offset = VDM_thrshPage * PAGE_SIZE + PAGE_OFFSET + THRSH_SPACING * thrshID
+                vdm_value_offset = vdm_valuePage * PAGE_SIZE + PAGE_OFFSET + VDM_SIZE * index
+                vdm_high_alarm_offset = vdm_thrshPage * PAGE_SIZE + PAGE_OFFSET + THRSH_SPACING * thrshID
                 vdm_low_alarm_offset = vdm_high_alarm_offset + 2
                 vdm_high_warn_offset = vdm_high_alarm_offset + 4
                 vdm_low_warn_offset = vdm_high_alarm_offset + 6
 
-                vdm_value_raw = self.xcvr_eeprom.read_flexible(vdm_value_offset, VDM_SIZE, True)
-                vdm_thrsh_high_alarm_raw = self.xcvr_eeprom.read_flexible(vdm_high_alarm_offset, VDM_SIZE, True)
-                vdm_thrsh_low_alarm_raw = self.xcvr_eeprom.read_flexible(vdm_low_alarm_offset, VDM_SIZE, True)
-                vdm_thrsh_high_warn_raw = self.xcvr_eeprom.read_flexible(vdm_high_warn_offset, VDM_SIZE, True)
-                vdm_thrsh_low_warn_raw = self.xcvr_eeprom.read_flexible(vdm_low_warn_offset, VDM_SIZE, True)
+                vdm_value_raw = self.xcvr_eeprom.read_raw(vdm_value_offset, VDM_SIZE, True)
+                vdm_thrsh_high_alarm_raw = self.xcvr_eeprom.read_raw(vdm_high_alarm_offset, VDM_SIZE, True)
+                vdm_thrsh_low_alarm_raw = self.xcvr_eeprom.read_raw(vdm_low_alarm_offset, VDM_SIZE, True)
+                vdm_thrsh_high_warn_raw = self.xcvr_eeprom.read_raw(vdm_high_warn_offset, VDM_SIZE, True)
+                vdm_thrsh_low_warn_raw = self.xcvr_eeprom.read_raw(vdm_low_warn_offset, VDM_SIZE, True)
                 if vdm_format == 'S16':
                     vdm_value = struct.unpack('>h',vdm_value_raw)[0] * scale
                     vdm_thrsh_high_alarm = struct.unpack('>h', vdm_thrsh_high_alarm_raw)[0] * scale
@@ -89,9 +89,9 @@ class CmisVdmApi(XcvrApi):
             vdm_high_warn_flag = bool((VDM_flag_page[vdm_flag_offset] >> (bit_offset+2)) & 0x1)
             vdm_low_warn_flag = bool((VDM_flag_page[vdm_flag_offset] >> (bit_offset+3)) & 0x1)
 
-            if vdm_type not in VDM_Page_data:
-                VDM_Page_data[vdm_type] = {
-                    VDM_lane[index]+1: [
+            if vdm_type not in vdm_Page_data:
+                vdm_Page_data[vdm_type] = {
+                    vdm_lane[index]+1: [
                         vdm_value,
                         vdm_thrsh_high_alarm,
                         vdm_thrsh_low_alarm,
@@ -104,7 +104,7 @@ class CmisVdmApi(XcvrApi):
                 }
 
             else:
-                VDM_Page_data[vdm_info_dict[0]][VDM_lane[index]+1] = [
+                vdm_Page_data[vdm_info_dict[0]][vdm_lane[index]+1] = [
                     vdm_value,
                     vdm_thrsh_high_alarm,
                     vdm_thrsh_low_alarm,
@@ -114,18 +114,18 @@ class CmisVdmApi(XcvrApi):
                     vdm_low_alarm_flag,
                     vdm_high_warn_flag,
                     vdm_low_warn_flag]
-        return VDM_Page_data
+        return vdm_Page_data
 
-    def get_VDM_allpage(self):
+    def get_vdm_allpage(self):
         vdm_page_supported_raw = self.xcvr_eeprom.read(consts.VDM_SUPPORTED_PAGE) & 0x3
         VDM_START_PAGE = 0x20
-        VDM = dict()
+        vdm = dict()
         self.xcvr_eeprom.write(consts.VDM_CONTROL, 128)
         time.sleep(5)
         self.xcvr_eeprom.write(consts.VDM_CONTROL, 0)
         time.sleep(1)
-        VDM_flag_page = self.xcvr_eeprom.read_flexible(VDM_FLAG_PAGE * PAGE_SIZE + PAGE_OFFSET, PAGE_SIZE)
+        vdm_flag_page = self.xcvr_eeprom.read_raw(VDM_FLAG_PAGE * PAGE_SIZE + PAGE_OFFSET, PAGE_SIZE)
         for page in range(VDM_START_PAGE, VDM_START_PAGE + vdm_page_supported_raw + 1):
-            VDM_current_page = self.get_VDM_page(page, VDM_flag_page)
-            VDM.update(VDM_current_page)
-        return VDM
+            vdm_current_page = self.get_vdm_page(page, vdm_flag_page)
+            vdm.update(vdm_current_page)
+        return vdm
