+++ sonic_platform_base/sonic_xcvr/mem_maps/public/cmis.py
@@ -26,21 +26,21 @@ class CmisMemMap(XcvrMemMap):
         super(CmisMemMap, self).__init__(codes)
 
         self.ADMIN_INFO = RegGroupField(consts.ADMIN_INFO_FIELD,
-            CodeRegField(consts.ID_FIELD, get_addr(0x0, 128), self.codes['sff8024'].XCVR_IDENTIFIERS),
+            CodeRegField(consts.ID_FIELD, get_addr(0x0, 128), self.codes.XCVR_IDENTIFIERS),
             StringRegField(consts.VENDOR_NAME_FIELD, get_addr(0x0, 129), size=16),
             HexRegField(consts.VENDOR_OUI_FIELD, get_addr(0x0, 145), size=3),
             StringRegField(consts.VENDOR_PART_NO_FIELD, get_addr(0x0, 148), size=16),
             StringRegField(consts.VENDOR_REV_FIELD, get_addr(0x0, 164), size=2),
             StringRegField(consts.VENDOR_SERIAL_NO_FIELD, get_addr(0x0, 166), size=16),
             StringRegField(consts.VENDOR_DATE_FIELD,get_addr(0x0,182), size = 8),
-            CodeRegField(consts.CONNECTOR_FIELD, get_addr(0x0, 203), self.codes['sff8024'].CONNECTORS),
-            CodeRegField(consts.HOST_ELECTRICAL_INTERFACE, get_addr(0x0, 86), self.codes['sff8024'].HOST_ELECTRICAL_INTERFACE),
-            CodeRegField(consts.MODULE_MEDIA_TYPE, get_addr(0x0, 85), self.codes['sff8024'].MODULE_MEDIA_TYPE),
-            CodeRegField(consts.MODULE_MEDIA_INTERFACE_850NM, get_addr(0x0, 87), self.codes['sff8024'].NM_850_MEDIA_INTERFACE),
-            CodeRegField(consts.MODULE_MEDIA_INTERFACE_SM, get_addr(0x0, 87), self.codes['sff8024'].SM_MEDIA_INTERFACE),
-            CodeRegField(consts.MODULE_MEDIA_INTERFACE_PASSIVE_COPPER, get_addr(0x0, 87), self.codes['sff8024'].PASSIVE_COPPER_MEDIA_INTERFACE),
-            CodeRegField(consts.MODULE_MEDIA_INTERFACE_ACTIVE_CABLE, get_addr(0x0, 87), self.codes['sff8024'].ACTIVE_CABLE_MEDIA_INTERFACE),
-            CodeRegField(consts.MODULE_MEDIA_INTERFACE_BASE_T, get_addr(0x0, 87), self.codes['sff8024'].BASE_T_MEDIA_INTERFACE),
+            CodeRegField(consts.CONNECTOR_FIELD, get_addr(0x0, 203), self.codes.CONNECTORS),
+            CodeRegField(consts.HOST_ELECTRICAL_INTERFACE, get_addr(0x0, 86), self.codes.HOST_ELECTRICAL_INTERFACE),
+            CodeRegField(consts.MODULE_MEDIA_TYPE, get_addr(0x0, 85), self.codes.MODULE_MEDIA_TYPE),
+            CodeRegField(consts.MODULE_MEDIA_INTERFACE_850NM, get_addr(0x0, 87), self.codes.NM_850_MEDIA_INTERFACE),
+            CodeRegField(consts.MODULE_MEDIA_INTERFACE_SM, get_addr(0x0, 87), self.codes.SM_MEDIA_INTERFACE),
+            CodeRegField(consts.MODULE_MEDIA_INTERFACE_PASSIVE_COPPER, get_addr(0x0, 87), self.codes.PASSIVE_COPPER_MEDIA_INTERFACE),
+            CodeRegField(consts.MODULE_MEDIA_INTERFACE_ACTIVE_CABLE, get_addr(0x0, 87), self.codes.ACTIVE_CABLE_MEDIA_INTERFACE),
+            CodeRegField(consts.MODULE_MEDIA_INTERFACE_BASE_T, get_addr(0x0, 87), self.codes.BASE_T_MEDIA_INTERFACE),
             NumberRegField(consts.LANE_COUNT, get_addr(0x0, 88), format="B", size=1),
             NumberRegField(consts.HOST_LANE_ASSIGNMENT_OPTION, get_addr(0x0, 89), format="B", size=1),
             NumberRegField(consts.MEDIA_LANE_ASSIGNMENT_OPTION, get_addr(0x1, 176), format="B", size=1),
@@ -52,7 +52,7 @@ class CmisMemMap(XcvrMemMap):
             NumberRegField(consts.ACTIVE_APSEL_HOSTLANE_6, get_addr(0x11, 211), format="B", size=1),
             NumberRegField(consts.ACTIVE_APSEL_HOSTLANE_7, get_addr(0x11, 212), format="B", size=1),
             NumberRegField(consts.ACTIVE_APSEL_HOSTLANE_8, get_addr(0x11, 213), format="B", size=1),
-            CodeRegField(consts.MEDIA_INTERFACE_TECH, get_addr(0x0, 212), self.codes['cmis_code'].MEDIA_INTERFACE_TECH),
+            CodeRegField(consts.MEDIA_INTERFACE_TECH, get_addr(0x0, 212), self.codes.MEDIA_INTERFACE_TECH),
             NumberRegField(consts.HW_MAJOR_REV, get_addr(0x1, 130), size=1),
             NumberRegField(consts.HW_MINOR_REV, get_addr(0x1, 131), size=1),
             NumberRegField(consts.CMIS_REVISION, get_addr(0x0, 1), format="B", size=1),
@@ -135,7 +135,7 @@ class CmisMemMap(XcvrMemMap):
             NumberRegField(consts.MODULE_FLAG_BYTE2, get_addr(0x0, 10), size=1),
             NumberRegField(consts.MODULE_FLAG_BYTE3, get_addr(0x0, 11), size=1),
             NumberRegField(consts.CDB1_STATUS, get_addr(0x0, 37), size=1),
-            CodeRegField(consts.MODULE_FAULT_CAUSE, get_addr(0x0, 41), self.codes['cmis_code'].MODULE_FAULT_CAUSE),
+            CodeRegField(consts.MODULE_FAULT_CAUSE, get_addr(0x0, 41), self.codes.MODULE_FAULT_CAUSE),
         )
 
         self.TRANS_LANE_STATUS = RegGroupField(consts.TRANS_LANE_STATUS_FIELD,
@@ -241,8 +241,5 @@ class CmisMemMap(XcvrMemMap):
             NumberRegField(consts.CDB_SEQ_WRITE_LENGTH_EXT, get_addr(0x01, 164), size=1),
             NumberRegField(consts.CDB_RPL_LENGTH, get_addr(0x9f, 134), size=1, ro=False),
             NumberRegField(consts.CDB_RPL_CHKCODE, get_addr(0x9f, 135), size=1, ro=False),
-            ListRegField(consts.CDB_LPL, get_addr(0x9f, 136), format='120B', size=120, ro=False),
-            ListRegField(consts.CDB_CMD, get_addr(0x9f, 128), format='2B', size=2, ro=False),
-            ListRegField(consts.CDB_WRITE_MSG, get_addr(0x9f, 130), format='126B', size=126, ro=False),
         )
         # TODO: add remaining fields
