+++ sonic_platform_base/sonic_xcvr/xcvr_eeprom.py
@@ -25,7 +25,7 @@ class XcvrEeprom(object):
       raw_data = self.reader(field.get_offset(), field.get_size())
       return field.decode(raw_data) if raw_data else None
 
-   def read_flexible(self, offset, size, return_raw = False):
+   def read_raw(self, offset, size, return_raw = False):
       """
       Read values from a field in EEPROM in a more flexible way
 
@@ -38,10 +38,10 @@ class XcvrEeprom(object):
          The value(s) of the field, if the read is successful and None otherwise
       """
       raw_data = self.reader(offset, size)
-      if return_raw:
-         return raw_data
       if raw_data is None:
          return None
+      if return_raw:
+         return raw_data
       else:
          if size == 1:
             data = struct.unpack("%dB" %size, raw_data)[0]
@@ -69,7 +69,7 @@ class XcvrEeprom(object):
          encoded_data = field.encode(value)
       return self.writer(field.get_offset(), field.get_size(), encoded_data)
 
-   def write_flexible(self, offset, size, bytearray_data):
+   def write_raw(self, offset, size, bytearray_data):
       """
       Write values to a field in EEPROM in a more flexible way
 
@@ -79,7 +79,7 @@ class XcvrEeprom(object):
 
          size: an integer indicating how many bytes to write to
 
-         bytearray_data: a bytearray as write bugger to be written into EEPROM
+         bytearray_data: a bytearray as write buffer to be written into EEPROM
 
       Returns:
          Boolean, True if the write is successful and False otherwise
